Added Auslastung back to the mensa; added the ViewModel to the toolbar; consinstency of naming (e.g. selectedCanteen to selectedCafeteria; extracted the MapScreen
